[feat/daengle-32] 빈 2개 이상 등록 에러로 인한 GlobalController auth 모듈에만 등록
