+++ daengle-auth/src/main/java/ddog/auth/exception/common/GlobalControllerAdvice.java
@@ -3,6 +3,7 @@ package ddog.auth.exception.common;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
@@ -38,6 +39,18 @@ public class GlobalControllerAdvice {
                         new CustomError(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR, null)
                 ),
                 HttpStatus.INTERNAL_SERVER_ERROR
+        );
+    }
+
+    @ExceptionHandler(HttpMessageNotReadableException.class)  //ENUM 값 불일치
+    public ResponseEntity<?> handleHttpMessageNotReadableException(HttpMessageNotReadableException e) {
+        return new ResponseEntity<>(
+                new CommonResponseEntity<>(
+                        false,
+                        null,
+                        new CustomError(e.getMessage(), HttpStatus.BAD_REQUEST, null)
+                ),
+                HttpStatus.BAD_REQUEST
         );
     }
 }
